+++ src/main/java/com/sparta/newsfeed/common/exception/post/NotFoundPostException.java
@@ -1,4 +1,4 @@
-package com.sparta.newsfeed.common.exception;
+package com.sparta.newsfeed.common.exception.post;
 
 public class NotFoundPostException extends RuntimeException {
     private static final String MESSAGE = "포스트를 찾을 수 없습니다.";
