+++ src/main/java/com/sparta/newsfeed/common/exception/user/NotFoundUserException.java
@@ -1,4 +1,4 @@
-package com.sparta.newsfeed.common.exception;
+package com.sparta.newsfeed.common.exception.user;
 
 public class NotFoundUserException extends RuntimeException {
     private static final String MESSAGE = "유저를 찾을 수 없습니다.";
